Create build config for Medianode-Q.
The i.MX6DL doesn't have enough oomph for our purposes, so we're swapping in
an i.MX6D (which for all intents and purposes is an i.MX6Q with two cores
missing).

Create a new config 'mx6qmedianode', add to boards.cfg.

WARNING: The DRAM configuration table is copied from
board/freescale/imx/ddr/mx6q_4x_mt41j128.cfg, which is an 800MHz setting.
This will almost certainly need to be tuned properly to the DRAMs we're
actually using and the clock speed we want.