Fix fw_{set,print}env to accept dev names >= 16 chars.
zOMG, this is awful stuff.  The buffer to contain the device name from which
to read/write U-Boot variables was 16 bytes in size.  This chopped off
exactly the last char in "/dev/mmcblk1boot0", which made it impossible for
us to open for writing.  Lord help you if you wanted to write to
/dev/disk/by-uuid/1ba3453f-5241-...

Also: Note how chopping one char off "/dev/mmcblk1boot0" would leave exactly
16 chars.  Where's the terminating NULL?  Why, in the next struct field,
'devoff'.  In other words, the sscanf() was writing beyond the end of the
output buffer.  This program is a pile of segfaults just waiting to happen.

Increase devname buffer from 16 to 256 bytes.  Increase line buffer from 128
to 512 bytes.  Convert sscanf() call to use auto-allocated temporary buffer
for the devname string ("%ms"), then strncpy() that string into the
DEVNAME() entry.