Boot kernel with read-only rootfs.
Alter mmcroot and mmcroot_eval variables to pass 'ro' option to the kernel,
causing the rootfs to be mounted read-only.  This only changes the
compile-time defaults -- nodes that have made changes to their U-Boot
variables still have the old value and will need to be explicitly updated.

Pursuant to XN-3774.