Add steering for Falcon builds.
U-Boot hard-codes the name of the .dtb file it's going to load alongside the
kernel.  Unfortunately, unlike the kernel (which has the generic name of
'uImage'), the DTB filename is target-specific, the result of inscrutable
transformations of machine, arch, and other names.

Add a new config option to the targets defined in boards.cfg --
CONFIG_MEDIANODE and CONFIG_FALCON.  Use these defines to select what DTB
filename to load from the boot media.

Pursuant to XN-5918.